Module google_billingbudgets1_beta1::api

Structs

BillingAccountBudgetCreateCall
Creates a new budget. See Quotas and limits for more information on the limits of the number of budgets you can create.

BillingAccountBudgetDeleteCall
Deletes a budget. Returns successfully if already deleted.

BillingAccountBudgetGetCall
Returns a budget. WARNING: There are some fields exposed on the Google Cloud Console that aren't available on this API. When reading from the API, you will not see these fields in the return value, though they may have been set in the Cloud Console.

BillingAccountBudgetListCall
Returns a list of budgets for a billing account. WARNING: There are some fields exposed on the Google Cloud Console that aren't available on this API. When reading from the API, you will not see these fields in the return value, though they may have been set in the Cloud Console.

BillingAccountBudgetPatchCall
Updates a budget and returns the updated budget. WARNING: There are some fields exposed on the Google Cloud Console that aren't available on this API. Budget fields that are not exposed in this API will not be changed by this method.

BillingAccountMethods
A builder providing access to all methods supported on billingAccount resources. It is not used directly, but through the CloudBillingBudget hub.

CloudBillingBudget
Central instance to access all CloudBillingBudget related resource activities

GoogleCloudBillingBudgetsV1beta1AllUpdatesRule
AllUpdatesRule defines notifications that are sent based on budget spend and thresholds.

GoogleCloudBillingBudgetsV1beta1Budget
A budget is a plan that describes what you expect to spend on Cloud projects, plus the rules to execute as spend is tracked against that plan, (for example, send an alert when 90% of the target spend is met). The budget time period is configurable, with options such as month (default), quarter, year, or custom time period.

GoogleCloudBillingBudgetsV1beta1BudgetAmount
The budgeted amount for each usage period.

GoogleCloudBillingBudgetsV1beta1CreateBudgetRequest
Request for CreateBudget

GoogleCloudBillingBudgetsV1beta1CustomPeriod
All date times begin at 12 AM US and Canadian Pacific Time (UTC-8).

GoogleCloudBillingBudgetsV1beta1Filter
A filter for a budget, limiting the scope of the cost to calculate.

GoogleCloudBillingBudgetsV1beta1LastPeriodAmount
Describes a budget amount targeted to last period's spend. At this time, the amount is automatically 100% of last period's spend; that is, there are no other options yet. Future configuration will be described here (for example, configuring a percentage of last period's spend).

GoogleCloudBillingBudgetsV1beta1ListBudgetsResponse
Response for ListBudgets

GoogleCloudBillingBudgetsV1beta1ThresholdRule
ThresholdRule contains a definition of a threshold which triggers an alert (a notification of a threshold being crossed) to be sent when spend goes above the specified amount. Alerts are automatically e-mailed to users with the Billing Account Administrator role or the Billing Account User role. The thresholds here have no effect on notifications sent to anything configured under Budget.all_updates_rule.

GoogleCloudBillingBudgetsV1beta1UpdateBudgetRequest
Request for UpdateBudget

GoogleProtobufEmpty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

GoogleTypeDate
Represents a whole or partial calendar date, such as a birthday. The time of day and time zone are either specified elsewhere or are insignificant. The date is relative to the Gregorian Calendar. This can represent one of the following: * A full date, with non-zero year, month, and day values * A month and day value, with a zero year, such as an anniversary * A year on its own, with zero month and day values * A year and month value, with a zero day, such as a credit card expiration date Related types are google.type.TimeOfDay and google.protobuf.Timestamp.

GoogleTypeMoney
Represents an amount of money with its currency type.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.